Added CommandUi to compartmentalize, finish up javadocs. Made some minor edit to Parser to fix LocalDate recognition.
