+++ src/main/java/herrekt/Ui.java
@@ -14,28 +14,6 @@ public class Ui {
         System.out.println("Bye. Hope you will complete everything for today!");
     }
 
-    void printNumberOfTasks(TaskList tasks) {
-        if (tasks.getSize() <= 1) {
-            System.out.println("For now, you have "
-                    + tasks.getSize()
-                    + " task on the list");
-        } else {
-            System.out.println("For now, you have "
-                    + tasks.getSize()
-                    + " tasks on the list");
-        }
-    }
-
-    void printMarkTaskAsDone(int taskNumber, TaskList tasks) {
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println("  " + tasks.getTask(taskNumber - 1).toString());
-    }
-
-    void printTaskAlreadyMarkedAsDone(int taskNumber, TaskList tasks) {
-        System.out.println("Task has already been marked as done.");
-        System.out.println("Task in question: " + tasks.getTask(taskNumber - 1).toString());
-    }
-
     void printTaskList(TaskList tasks) {
         if (tasks.getSize() == 0) {
             System.out.println("You have no outstanding tasks. Good Job!");
@@ -43,25 +21,6 @@ public class Ui {
             System.out.println("Here are the tasks in your list:" + "\n" + tasks.toString());
         }
     }
-
-    void printMatchingTaskList(TaskList tasks) {
-        if (tasks.getSize() == 0) {
-            System.out.println("Couldn't find any matches");
-        } else {
-            System.out.println("Here are the matching tasks in your list:" + "\n" + tasks.toString());
-        }
-    }
-
-    void printTaskDeleted(int taskNumber, TaskList tasks) {
-        int index = taskNumber - 1;
-        System.out.println("Alright. I've removed this task: "
-                + "\n" + "  " + tasks.getTask(index).toString());
-        System.out.println("Now you have "
-                + (tasks.getSize() - 1)
-                + " tasks left to do");
-    }
-
-
 
     void printLoadingError() {
         System.out.println("ERROR! Could not locate save.txt.");
